Ecorp-formapi-main/requirements.txt
azure-ai-formrecognizer==3.2.1
azure-common==1.1.28
azure-core==1.26.4
docx==0.2.4

